clean and rename to private
