+++ cpp-terminal/CMakeLists.txt
@@ -1,10 +1,10 @@
 # configure version information
 configure_file(version.cpp.in version.cpp)
 
-add_subdirectory(platforms)
+add_subdirectory(private)
 
 # create and configure library target
-add_library(cpp-terminal terminal_initializer.cpp iostream_initializer.cpp mouse.cpp args.cpp focus.cpp buffer.cpp iostream.cpp stream.cpp prompt.cpp window.cpp terminal.cpp color.cpp key.cpp event.cpp screen.cpp options.cpp cursor.cpp style.cpp "${CMAKE_CURRENT_BINARY_DIR}/version.cpp")
+add_library(cpp-terminal terminal_impl.cpp terminal_initializer.cpp iostream_initializer.cpp mouse.cpp args.cpp focus.cpp buffer.cpp iostream.cpp stream.cpp prompt.cpp window.cpp terminal.cpp color.cpp key.cpp event.cpp screen.cpp options.cpp cursor.cpp style.cpp "${CMAKE_CURRENT_BINARY_DIR}/version.cpp")
 target_link_libraries(cpp-terminal PRIVATE Warnings::Warnings cpp-terminal::cpp-terminal-platforms)
 target_compile_options(cpp-terminal PRIVATE $<$<CXX_COMPILER_ID:MSVC>:/utf-8>)
 target_include_directories(cpp-terminal PUBLIC $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}> $<BUILD_INTERFACE:${PROJECT_BINARY_DIR}> $<INSTALL_INTERFACE:include>)
