+++ cpp-terminal/private/README.md
@@ -1,0 +1,1 @@
+# Private folder for cpp-terminal
