+++ cpp-terminal/private/conversion.cpp
@@ -7,7 +7,7 @@
 * SPDX-License-Identifier: MIT
 */
 
-#include "cpp-terminal/platforms/conversion.hpp"
+#include "cpp-terminal/private/conversion.hpp"
 
 #include "cpp-terminal/exception.hpp"
