+++ cpp-terminal/private/conversion.hpp
