+++ cpp-terminal/private/exception.cpp
@@ -7,10 +7,10 @@
 * SPDX-License-Identifier: MIT
 */
 
-#include "cpp-terminal/platforms/exception.hpp"
+#include "cpp-terminal/private/exception.hpp"
 
 #if defined(_WIN32)
-  #include "cpp-terminal/platforms/unicode.hpp"
+  #include "cpp-terminal/private/unicode.hpp"
 
   #include <memory>
   #include <windows.h>
