+++ cpp-terminal/private/exception.hpp
