+++ cpp-terminal/private/macros.hpp
