+++ cpp-terminal/window.hpp
@@ -10,7 +10,6 @@
 #pragma once
 
 #include "cpp-terminal/color.hpp"
-#include "cpp-terminal/platforms/macros.hpp"
 #include "cpp-terminal/style.hpp"
 
 #include <vector>
@@ -53,9 +52,9 @@ private:
 public:
   Window(const std::size_t& w, const std::size_t& h) : w{w}, h{h}, chars(w * h, ' '), m_fg(w * h, {0, 0, 0}), m_bg(w * h, {0, 0, 0}), m_fg_reset(w * h, true), m_bg_reset(w * h, true), m_style(w * h, Style::Reset){};
 
-  CPP_TERMINAL_NODISCARD std::size_t get_w() const;
+  std::size_t get_w() const;
 
-  CPP_TERMINAL_NODISCARD std::size_t get_h() const;
+  std::size_t get_h() const;
 
   void set_char(const std::size_t&, const std::size_t&, const char32_t&);
