+++ tests/test_terminal.test.cpp
@@ -8,7 +8,7 @@
 */
 
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "cpp-terminal/platforms/conversion.hpp"
+#include "cpp-terminal/private/conversion.hpp"
 #include "doctest/doctest.h"
 
 TEST_CASE("UTF conversions")
